refactor: 코드정리/주석 및 사용하지 않는 import 삭제
